+++ tournament/views_main.py
@@ -3,7 +3,7 @@ from passlib.hash import pbkdf2_sha256
 
 from tournament import app, login_manager, flask_login
 from tournament.forms import Register, Login
-from tournament.models import User
+from tournament.models_user import User
 
 
 @app.route('/')
